# Conflicts:
#	Du_Toan_Xay_Dung/Scripts/dutoan/Estimate_Work/services/dataService.js

+++ Du_Toan_Xay_Dung/Controllers/CongTrinhController.cs
@@ -6,7 +6,7 @@ using System.Web.Mvc;
 using Du_Toan_Xay_Dung.Models;
 using Du_Toan_Xay_Dung.Handlers;
 using System.Web.UI.WebControls;
-using System.IO; 
+using System.IO;
 using System.Web.UI;
 //using CRUDDeom.Models;
 using OfficeOpenXml;
@@ -23,13 +23,14 @@ namespace Du_Toan_Xay_Dung.Controllers
         [PageLogin]
         public ActionResult Index()
         {
-            ////var model = ListAllPageging(page, pagesize);
-            
-            //ViewData["List_CongTrinh"] = _db.Buildings.Where(i => i.Email.Equals(SessionHandler.User.Email)).Select(i => new BuildingViewModel(i)).ToList();
-            ////ViewData["List_CongTrinh_Null"] = _db.Buildings.Where(i => i.Email.Equals(SessionHandler.User.Email) && !i.BuildingItems.Any(o => o.ID.Equals(i.ID))).Select(i => i.ID).ToList();
-            ////ViewData["list_hinhanh"] = _db.Images_Urls.Select(i => new Images_CongTrinhViewModel(i)).ToList();
-
             return View();
+        }
+
+        [PageLogin]
+        public JsonResult Get_AllInfoBuildings()
+        {
+            var list = _db.Buildings.Where(i => i.Email.Equals(SessionHandler.User.Email)).OrderByDescending(i => i.ID).Select(i => new BuildingViewModel(i)).ToList();
+            return Json(list, JsonRequestBehavior.AllowGet);
         }
 
         [PageLogin]
@@ -42,37 +43,40 @@ namespace Du_Toan_Xay_Dung.Controllers
         public ActionResult BuildingItem()
         {
             ViewData["List_CongTrinh"] = _db.Buildings.Where(i => i.Email.Equals(SessionHandler.User.Email)).Select(i => new BuildingViewModel(i)).ToList();
-            //ViewData["List_HangMuc"] = _db.BuildingItems.Select(i => new BuildingItemViewModel(i)).ToList();
-
             return View();
         }
 
         public JsonResult Get_HangMuc(int id)
         {
-            //var list = _db.Buildings.Join(_db.BuildingItems, bid => bid.ID, biid => biid.Building_ID, (bid, biid) => new
-            //{
-            //    Building = bid,
-            //    BuildingItem = biid
-            //}).Where(i => i.Building.Email.Equals(SessionHandler.User.Email)).Select(i => new
-            //{
-            //    ID = i.BuildingItem.ID,
-            //    Building_ID = i.BuildingItem.Building_ID,
-            //    Name = i.BuildingItem.Name,
-            //    Description = i.BuildingItem.Description,
-            //    Sum = i.BuildingItem.Sum
-            //}).ToList();
-
             var list = _db.BuildingItems.Where(i => i.Building_ID.Equals(id)).Select(i => new BuildingItemViewModel(i)).ToList();
 
             return Json(list, JsonRequestBehavior.AllowGet);
         }
 
         [PageLogin]
-        public JsonResult Get_AllInfoBuildings()
+        public JsonResult get_dataSavetoDropbox(string id)
         {
-            var list = _db.Buildings.Where(i => i.Email.Equals(SessionHandler.User.Email)).Select(i => new BuildingViewModel(i)).ToList();
-            return Json(list, JsonRequestBehavior.AllowGet);
+            var list_temp = _db.Images_Urls.Where(i => i.Building_ID.Equals(id)).Select(i => new Images_CongTrinhViewModel(i)).ToList();
+
+            List<Images_CongTrinhViewModel> list_images = new List<Images_CongTrinhViewModel>();
+
+            foreach (var item in list_temp)
+            {
+                string path = Server.MapPath(@"~" + item.Url);
+
+                byte[] imageArray = System.IO.File.ReadAllBytes(path);
+                string base64ImageRepresentation = Convert.ToBase64String(imageArray);
+
+                var Arr_temp = item.Url.Split('/');
+                var Arr_name = Arr_temp[4].Split('.');
+
+                list_images.Add(new Images_CongTrinhViewModel() { name = Arr_temp[4], src = "data:image/" + Arr_name[1].ToLower() + ";base64," + base64ImageRepresentation });
+
+            }
+
+            return Json(list_images, JsonRequestBehavior.AllowGet);
         }
+
 
         [PageLogin]
         [HttpPost]
@@ -227,11 +231,15 @@ namespace Du_Toan_Xay_Dung.Controllers
 
                     string location = Server.MapPath(@"~/Images/CongTrinh/Building" + Convert.ToString(building_id) + "/");
                     DirectoryInfo directory = new DirectoryInfo(location);
-                    foreach (FileInfo file in directory.GetFiles())
+
+                    if (directory.Exists)
                     {
-                        file.Delete();
+                        foreach (FileInfo file in directory.GetFiles())
+                        {
+                            file.Delete();
+                        }
+                        directory.Delete(true);
                     }
-                    directory.Delete(true);
 
                     if (hangmuc.Count != 0)
                     {
